+++ griptape/chunkers/base_chunker.py
@@ -21,6 +21,11 @@ class BaseChunker(ABC):
         default=Factory(lambda self: self.tokenizer.max_input_tokens, takes_self=True), kw_only=True
     )
 
+    @max_tokens.validator  # pyright: ignore
+    def validate_max_tokens(self, _, max_tokens: int) -> None:
+        if max_tokens < 0:
+            raise ValueError("max_tokens must be 0 or greater.")
+
     def chunk(self, text: TextArtifact | str) -> list[TextArtifact]:
         text = text.value if isinstance(text, TextArtifact) else text