+++ tests/unit/engines/summary/test_prompt_summary_engine.py
@@ -1,6 +1,7 @@
 import pytest
 from griptape.artifacts import TextArtifact, ListArtifact
 from griptape.engines import PromptSummaryEngine
+from griptape.utils import PromptStack
 from tests.mocks.mock_prompt_driver import MockPromptDriver
 import os
 
@@ -26,11 +27,16 @@ class TestPromptSummaryEngine:
             PromptSummaryEngine(prompt_driver=MockPromptDriver(), max_token_multiplier=10000)
 
     def test_chunked_summary(self, engine):
-        engine = PromptSummaryEngine(prompt_driver=MockPromptDriver(), max_token_multiplier=0.03)
+        def smaller_input(prompt_stack: PromptStack):
+            return prompt_stack.inputs[0].content[: (len(prompt_stack.inputs[0].content) // 2)]
+
+        engine = PromptSummaryEngine(prompt_driver=MockPromptDriver(mock_output="smaller_input"))
 
         def copy_test_resource(resource_path: str):
             file_dir = os.path.dirname(__file__)
             full_path = os.path.join(file_dir, "../../../resources", resource_path)
             full_path = os.path.normpath(full_path)
+            with open(full_path) as f:
+                return f.read()
 
-        assert engine.summarize_text(copy_test_resource("test.txt"))
+        assert engine.summarize_text(copy_test_resource("test.txt") * 50)